--- src/main/java/com/jjs/ClothingInventorySaleReformPlatform/domain/reform/ProgressStatus.java
@@ -1,6 +1,0 @@
-package com.jjs.ClothingInventorySaleReformPlatform.domain.reform;
-
-public enum ProgressStatus {
-
-    REFORM_START, REFORMING, REFORM_COMPLETE // 리폼 시작, 리폼 중, 리폼 완료
-}
