+++ src/main/java/com/jjs/ClothingInventorySaleReformPlatform/domain/reform/progressmanagement/ProgressStatus.java
@@ -1,0 +1,6 @@
+package com.jjs.ClothingInventorySaleReformPlatform.domain.reform.progressmanagement;
+
+public enum ProgressStatus {
+
+    REFORM_START, REFORMING, REFORM_COMPLETE // 리폼 시작, 리폼 중, 리폼 완료
+}
