+++ src/main/java/com/jjs/ClothingInventorySaleReformPlatform/repository/reform/estimate/EstimateRepository.java
@@ -1,0 +1,10 @@
+package com.jjs.ClothingInventorySaleReformPlatform.repository.reform.estimate;
+
+import com.jjs.ClothingInventorySaleReformPlatform.domain.reform.estimate.Estimate;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface EstimateRepository extends JpaRepository<Estimate, Long> {
+    Optional<Estimate> findEstimateById(Long id);
+}
